// src/utils/enums.ts
export enum ETranslateFR {
  REQUIRED_PASSWORD = 'Password is required',
  REQUIRED_EMAIL = 'Email is required',
  REQUIRED_LASTNAME = 'LastName is required',
  REQUIRED_FIRSTNAME = 'Gender is required',
  REQUIRED_GENDER = 'FirstName is required',
  REQUIRED_DATE = 'BirthDay is required',
  REQUIRED_ADRESSE = 'Adress is required',
  ERROR_REGEX_PASSWORD = 'Must include at least 1 upper ,1 lower letter, 1 number and 1 special character',
  ERROR_REGEX_EMAIL = 'Email must be valid',
  ERROR_REGEX_NAME = 'Name must be valid',
  ERROR_REGEX_ADRESSE = 'Adress must be valid',
  ERROR_REGEX_LASTNAME = 'LastName must be valid',
  CONFIRM_PASSWORD = 'You must confirm your password',
  REGEX_CONFIRM_PASSWORD = 'Password and confirmation do not match',
  REQUIRED_PHONE = 'Phone number is required',
  REQUIRED = 'Required',
  ERROR_DATE = 'End Date is required',
  IS_AFTER = 'is after start',
  END_DATE = 'End Date must be after Start Date',
  STRONGER_PASSWORD = 'Please enter a stronger password',
  NEW_PASSWORD = 'newPassword',
  MATCH_PASSWORD = 'Password must match',
  DIGIT_NUMBER = 'Phone number must contain 8 digits',
}